SDL QObject generator warnings reduced Test Data updated
