+++ Include/imtsdlgencpp/Test/TestData/ReferenceData/ComplexCollectionFilterTest/SDL/1.0/CPP/ComplexCollectionFilter.cpp
@@ -3530,8 +3530,8 @@ void CComplexCollectionFilterObject::SetDistinctFields(const QVariant& v)
 		CComplexCollectionFilter::distinctFields->clear();
 	}
 	for (const auto& tempValue: v.value<QList<QString>>()){
-		istd::TNullableValue<QByteArray> value(tempValue.toUtf8());
-		CComplexCollectionFilter::distinctFields->append(value);
+		istd::TNullableValue<QByteArray> tempItem(tempValue.toUtf8());
+		CComplexCollectionFilter::distinctFields->append(tempItem);
 	}
 
 	distinctFieldsChanged();
