+++ Include/imtsdlgencpp/Test/TestData/ReferenceData/Printer/SDL/1.0/CPP/DigitalPrinter.h
@@ -33,6 +33,40 @@ namespace sdl::V1_0::modsdl
 {
 
 
+inline namespace DigitalPrinterSdlEnums
+{
+
+
+
+Q_NAMESPACE
+
+enum class PrintingTechnology {
+	Inkjet,
+	Toner,
+};
+
+Q_ENUM_NS(PrintingTechnology)
+
+
+class EnumPrintingTechnology: public QObject
+{
+	Q_OBJECT
+	Q_PROPERTY(QString Inkjet READ GetInkjet NOTIFY InkjetChanged)
+	Q_PROPERTY(QString Toner READ GetToner NOTIFY TonerChanged)
+
+protected:
+	static QString GetInkjet() { return "Inkjet"; }
+	static QString GetToner() { return "Toner"; }
+
+signals:
+	void InkjetChanged();
+	void TonerChanged();
+};
+
+
+
+} // inline namespace DigitalPrinterSdlEnums
+
 
 class CDigitalPrinterSpecification
 {
