+++ Include/imtsdlgencpp/Test/TestData/ReferenceData/Printer/SDL/1.0/CPP/DigitalPrinter_fwd.h
@@ -1,7 +1,6 @@
 #pragma once
 
 
-#include <QtCore/QObject>
 
 
 // custom types includes
@@ -14,30 +13,8 @@ namespace sdl::V1_0::modsdl
 inline namespace DigitalPrinterSdlEnums
 {
 
-Q_NAMESPACE
 
-enum class PrintingTechnology {
-	Inkjet,
-	Toner,
-};
-
-Q_ENUM_NS(PrintingTechnology)
-
-
-class EnumPrintingTechnology: public QObject
-{
-	Q_OBJECT
-	Q_PROPERTY(QString Inkjet READ GetInkjet NOTIFY InkjetChanged)
-	Q_PROPERTY(QString Toner READ GetToner NOTIFY TonerChanged)
-
-protected:
-	static QString GetInkjet() { return "Inkjet"; }
-	static QString GetToner() { return "Toner"; }
-
-signals:
-	void InkjetChanged();
-	void TonerChanged();
-};
+enum class PrintingTechnology;
 
 
 } // inline namespace DigitalPrinterSdlEnums
